requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.33.0
httpx==0.28.1
python-statemachine==2.5.0
PyYAML==6.0.3
Jinja2==3.1.6
pycdlib==1.14.0
SQLAlchemy==2.0.39
aiosqlite==0.21.0
